Use script-codec
