+++ packages/web3/src/codec/script-codec.test.ts
@@ -16,10 +16,24 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
+import {
+  AddressConst,
+  AssertWithErrorCode,
+  ByteConst,
+  CallExternal,
+  LoadLocal,
+  U256Const0,
+  U256Const1,
+  U256Eq,
+  StoreLocal
+} from './instr-codec'
 import { Buffer } from 'buffer/'
-import { web3, Fields, FieldsSig, buildScriptByteCode } from '@alephium/web3'
-import { randomContractId } from '@alephium/web3-test'
+import { Script, bs58, web3, Fields, FieldsSig, buildScriptByteCode } from '@alephium/web3'
+import { randomContractId, testAddress } from '@alephium/web3-test'
 import { scriptCodec } from './script-codec'
+import { Method } from './method-codec'
+import { LockupScript } from './lockup-script-codec'
+import { DestroyAdd, GreeterMain } from '../../../../artifacts/ts'
 
 describe('Encode & decode scripts', function () {
   beforeAll(() => {
@@ -144,6 +158,64 @@ describe('Encode & decode scripts', function () {
     )
   })
 
+  it('should encode & decode TxScript from the project', () => {
+    const contractId = randomContractId()
+    const decodedTestAddress = Buffer.from(bs58.decode(testAddress))
+    const lockupScript = {
+      scriptType: decodedTestAddress[0],
+      script: {
+        publicKeyHash: decodedTestAddress.slice(1)
+      }
+    } as LockupScript
+    const contractIdByteString = {
+      length: { mode: 64, rest: Buffer.from(['32']) },
+      value: Buffer.from(contractId, 'hex')
+    }
+
+    testScript(DestroyAdd.script, { add: contractId, caller: testAddress }, [
+      {
+        isPublic: true,
+        assetModifier: 3,
+        argsLength: 0,
+        localsLength: 0,
+        returnLength: 0,
+        instrs: [AddressConst(lockupScript), U256Const1, U256Const0, ByteConst(contractIdByteString), CallExternal(3)]
+      }
+    ])
+
+    testScript(GreeterMain.script, { greeterContractId: contractId }, [
+      {
+        isPublic: true,
+        assetModifier: 3,
+        argsLength: 0,
+        localsLength: 2,
+        returnLength: 0,
+        instrs: [
+          ByteConst(contractIdByteString),
+          StoreLocal(0),
+          U256Const0,
+          U256Const1,
+          LoadLocal(0),
+          CallExternal(0),
+          U256Const1,
+          U256Eq,
+          U256Const0,
+          AssertWithErrorCode,
+          ByteConst(contractIdByteString),
+          StoreLocal(1),
+          U256Const0,
+          U256Const1,
+          LoadLocal(1),
+          CallExternal(0),
+          U256Const1,
+          U256Eq,
+          U256Const0,
+          AssertWithErrorCode
+        ]
+      }
+    ])
+  })
+
   async function testScriptCode(
     scriptCode: string,
     fields: Fields = {},
@@ -155,5 +227,22 @@ describe('Encode & decode scripts', function () {
     const decoded = scriptCodec.decode(Buffer.from(scriptBytecode, 'hex'))
     const encoded = scriptCodec.encode(decoded)
     expect(scriptBytecode).toEqual(encoded.toString('hex'))
+  }
+
+  function testScript(script: Script, fields: Fields, methods: Method[]) {
+    const txScriptBytecode = script.buildByteCodeToDeploy(fields)
+    const decodedTxScript = scriptCodec.decodeScript(Buffer.from(txScriptBytecode, 'hex'))
+
+    expect(scriptCodec.encodeScript(decodedTxScript).toString('hex')).toEqual(txScriptBytecode)
+
+    expect(decodedTxScript.methods.length).toEqual(methods.length)
+    decodedTxScript.methods.map((decodedMethod, index) => {
+      expect(decodedMethod.isPublic).toEqual(methods[index].isPublic)
+      expect(decodedMethod.assetModifier).toEqual(methods[index].assetModifier)
+      expect(decodedMethod.argsLength).toEqual(methods[index].argsLength)
+      expect(decodedMethod.localsLength).toEqual(methods[index].localsLength)
+      expect(decodedMethod.returnLength).toEqual(methods[index].returnLength)
+      expect(decodedMethod.instrs).toEqual(methods[index].instrs)
+    })
   }
 })
